pyproject.toml
[tool.poetry]
name = "tap-dashboard-reports"
version = "0.0.1"
description = "`tap-dashboard-reports` is a Singer tap for DasboardReports, built with the Meltano SDK for Singer Taps."
authors = ["Alexey Kvitinsy"]
keywords = [
    "ELT",
    "DasboardReports",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.11,>=3.7.1"
requests = "^2.25.1"
singer-sdk = "^0.4.4"
jinja2 = "^3.1.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
black = "^21.9b0"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "tap_dashboard_reports"

[build-system]
# Uncomment the pinned version in favor of the git URL once
# https://github.com/python-poetry/poetry-core/pull/257 is merged
# and a new poetry-core 1.0.x is released
# requires = ["poetry-core>=1.0.0"]
requires = ["poetry-core @ git+https://github.com/python-poetry/poetry-core.git@master"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-dashboard-reports = 'tap_dashboard_reports.tap:TapDashboardReports.cli'
